requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
install==1.3.5
Markdown==3.5.2
mccabe==0.7.0
packaging==24.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
